requirements.txt
# ─────────────────────────────────────────────────
# HadithiAI Live — Python Dependencies
# ─────────────────────────────────────────────────

# ── Web Framework & Server ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.19

# ── Google Cloud ──
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.74.0
google-cloud-logging==3.11.3
google-genai==1.1.0

# ── Configuration ──
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# ── Async Utilities ──
aiohttp==3.11.11
aiofiles==24.1.0

# ── Observability ──
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-instrumentation-fastapi==0.50b0
